Merge branch 'feature_64BitMigrationOrganized' into develop

+++ coast/foundation/io/Resolver.cpp
@@ -10,6 +10,7 @@
 #include "SystemLog.h"
 #include "Socket.h"
 #include "Tracer.h"
+#include <stdint.h>
 #if !defined(WIN32)
 #include <arpa/inet.h>		// for inet_ntop gethostbyname etc.
 #include <netdb.h>			// for gethostbyname etc.
@@ -67,7 +68,7 @@ public:
 	ROAnything getAliases() const { return fAliases; }
 
 	bool DNS2IP(String &ipAddress, const String &dnsName);
-	bool IP2DNS(const String &ipAddress, unsigned long addr);
+	bool IP2DNS(const String &ipAddress, uint32_t addr);
 };
 
 String Resolver::DNS2IPAddress( const String &dnsName, const String &dflt )
@@ -95,7 +96,7 @@ String Resolver::DNS2IPAddress( const String &dnsName, const String &dflt )
 String Resolver::IPAddress2DNS( const String &ipAddress, const String &dflt )
 {
 	StartTrace1(Resolver.IPAddress2DNS, "ip [" << ipAddress << "]");
-	unsigned long addr = EndPoint::MakeInetAddr(ipAddress);
+	uint32_t addr = EndPoint::MakeInetAddr(ipAddress);
 
 	SystemSpecific(Resolver) sysResolver;
 
@@ -129,7 +130,7 @@ bool SunResolver::DNS2IP(String &ipAddress, const String &dnsName)
 	return false;
 }
 
-bool SunResolver::IP2DNS(const String &ipAddress, unsigned long addr)
+bool SunResolver::IP2DNS(const String &ipAddress, uint32_t addr)
 {
 	StartTrace1(Resolver.IP2DNS, "<sun> ip [" << ipAddress << "]");
 	struct hostent result = { 0 };
@@ -164,7 +165,7 @@ bool Win32Resolver::DNS2IP(String &ipAddress, const String &dnsName)
 	return false;
 }
 
-bool Win32Resolver::IP2DNS(const String &ipAddress, unsigned long addr)
+bool Win32Resolver::IP2DNS(const String &ipAddress, uint32_t addr)
 {
 	struct hostent *hePtr;
 	if (hePtr = gethostbyaddr((char *)&addr, sizeof(addr), AF_INET)) {
@@ -191,7 +192,7 @@ bool AppleResolver::DNS2IP(String &ipAddress, const String &dnsName)
 	return false;
 }
 
-bool AppleResolver::IP2DNS(const String &ipAddress, unsigned long addr)
+bool AppleResolver::IP2DNS(const String &ipAddress, uint32_t addr)
 {
 	struct hostent *ptrHe;
 	int error_num = 0;
@@ -228,7 +229,7 @@ bool LinuxResolver::DNS2IP(String &ipAddress, const String &dnsName)
 	return false;
 }
 
-bool LinuxResolver::IP2DNS(const String &ipAddress, unsigned long addr)
+bool LinuxResolver::IP2DNS(const String &ipAddress, uint32_t addr)
 {
 	StartTrace1(Resolver.IP2DNS, "<linux> ip [" << ipAddress << "]");
 	struct hostent he;
